legal_service_quotes_microservice/requirements.txt
FastAPI
SQLAlchemy
uvicorn
pydantic
httpx
python-dotenv
openai
email-validator
pymysql
cryptography